+++ app/dashboard/account/actions.ts
@@ -116,3 +116,40 @@ export const handleStripeConnect = async () => {
 
     }
   };
+
+export const resetPasswordAction = async (formData: FormData) => {
+    const supabase = await createClient();
+  
+    const password = formData.get("password") as string;
+    const confirmPassword = formData.get("confirmPassword") as string;
+  
+    if (!password || !confirmPassword) {
+      encodedRedirect(
+        "error",
+        "/protected/reset-password",
+        "Password and confirm password are required",
+      );
+    }
+  
+    if (password !== confirmPassword) {
+      encodedRedirect(
+        "error",
+        "/protected/reset-password",
+        "Passwords do not match",
+      );
+    }
+  
+    const { error } = await supabase.auth.updateUser({
+      password: password,
+    });
+  
+    if (error) {
+      encodedRedirect(
+        "error",
+        "/protected/reset-password",
+        "Password update failed",
+      );
+    }
+  
+    encodedRedirect("success", "/protected/reset-password", "Password updated");
+  };
