+++ utils/storage.ts
@@ -1,21 +1,3 @@
-export async function uploadImage(file: File): Promise<string> {
-  const formData = new FormData();
-  formData.append('file', file);
-
-  const response = await fetch('/api/upload', {
-    method: 'POST',
-    body: formData,
-  });
-
-  if (!response.ok) {
-    const error = await response.json();
-    throw new Error(error.error || 'Failed to upload image');
-  }
-
-  const data = await response.json();
-  return data.url;
-} 
-
 export async function uploadAvatar(file: File): Promise<string> {
   const formData = new FormData();
   formData.append('file', file);
